+++ LightServer/Models/SetLightRequest.cs
@@ -5,15 +5,15 @@ public class SetLightRequest
     /// <summary>
     /// Номер порта: 1-8
     /// </summary>
-    public int Port { get; set; }
+    public Byte Port { get; set; }
     
     /// <summary>
     /// Яркость: 0-100
     /// </summary>
-    public int Brightness { get; set; }
+    public UInt16 Brightness { get; set; }
     
     /// <summary>
     /// Длительность вспышки в мс (0 = постоянно)
     /// </summary>
-    public int DurationMs { get; set; } = 0;
+    public UInt16 DurationMs { get; set; } = 0;
 }
